+++ AndroidSekmeliArayuz/app/src/main/java/com/yenitsrm/androidsekmeliarayuz/tabs/TabFragment1.java
@@ -1,4 +1,4 @@
-package com.yenitsrm.androidsekmeliarayuz;
+package com.yenitsrm.androidsekmeliarayuz.tabs;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -11,13 +11,21 @@ import java.util.ArrayList;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.GridLayoutManager;
 
-public class TabFragment2 extends Fragment {
+import com.yenitsrm.androidsekmeliarayuz.R;
+import com.yenitsrm.androidsekmeliarayuz.gallery.GalleryAdapter;
+import com.yenitsrm.androidsekmeliarayuz.gallery.ImageModel;
+
+public class TabFragment1 extends Fragment {
 
     GalleryAdapter mAdapter;
     RecyclerView mRecyclerView;
 
     ArrayList<ImageModel> data = new ArrayList<>();
 
+
+
+
+    //galeride gösterilecek rerimlerin url dizisi
     public static String IMGS[] = {
 
             "https://images.unsplash.com/photo-1444090542259-0af8fa96557e?q=80&fm=jpg&w=1080&fit=max&s=4b703b77b42e067f949d14581f35019b",
@@ -34,17 +42,22 @@ public class TabFragment2 extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
 
-        View view = inflater.inflate(R.layout.tab_fragment_2, container, false);
 
+        View view = inflater.inflate(R.layout.tab_fragment_1, container, false);
+
+
+        //url'lerden resimlerimizi üretiyoruz
         for (int i = 0; i < IMGS.length; i++) {
 
             ImageModel imageModel = new ImageModel();
             imageModel.setName("Image " + i);
             imageModel.setUrl(IMGS[i]);
             data.add(imageModel);
+
         }
 
-        mRecyclerView = (RecyclerView) view.findViewById(R.id.image_list_2);
+        //resimler galeri şeklinde görüntüleniyor
+        mRecyclerView = (RecyclerView) view.findViewById(R.id.image_list);
         mRecyclerView.setLayoutManager(new GridLayoutManager(getActivity().getBaseContext(), 3));
         mRecyclerView.setHasFixedSize(true); // Helps improve performance
         mAdapter = new GalleryAdapter(getActivity().getBaseContext(), data);
@@ -52,5 +65,4 @@ public class TabFragment2 extends Fragment {
 
         return view;
     }
-
 }
