+++ AndroidSekmeliArayuz/app/src/main/java/com/yenitsrm/androidsekmeliarayuz/tabs/TabFragment4.java
@@ -1,15 +1,17 @@
-package com.yenitsrm.androidsekmeliarayuz;
+package com.yenitsrm.androidsekmeliarayuz.tabs;
  
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
- 
-public class TabFragment3 extends Fragment {
+
+import com.yenitsrm.androidsekmeliarayuz.R;
+
+public class TabFragment4 extends Fragment {
  
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.tab_fragment_3, container, false);
+        return inflater.inflate(R.layout.tab_fragment_4, container, false);
     }
 }
